Validation Grades && Classrooms

+++ app/Http/Controllers/Grades/GradeController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Grades;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\GradeRequest;
 use App\Http\Requests\UpdateGradeRequest;
+use App\Models\Classrooms\Classroom;
 use App\Models\Grades\Grade;
 use Illuminate\Http\Request;
 
@@ -66,10 +67,15 @@ class GradeController extends Controller
         }
     }
 
-
     public function destroy(Request $request)
     {
-        try {
+        // check if are there any classes in this grade
+        $myClass_id = Classroom::where('grade_id', $request->id)->pluck('grade_id');
+
+        // if there are no classes in this grade ==  delete the grade
+
+        if ($myClass_id->count() == 0) {
+
             $grade = Grade::findOrFail($request->id);
 
             $grade->delete();
@@ -77,8 +83,13 @@ class GradeController extends Controller
             toastr()->success(__('messages.Delete'));
 
             return redirect()->route('grades.index');
-        } catch (\Exception $e) {
-            return redirect()->back()->withErrors(['error' => $e->getMessage()]);
+        }
+        // if there are classes in this grade alert error and redirect
+        else {
+
+            toastr()->error(__('grades.delete_Grade_Error'));
+
+            return redirect()->route('grades.index');
         }
     }
 }
